chat-server/requirements.txt
flask
Flask-SQLAlchemy
requests
python-dateutil
flask-cors
gevent
pymysql
